Add support for examining structs for query helpers.
pan can now determine the name of the SQL column associated with a
struct's property. It also created a "sqlTableNamer" interface that
allows types to specify the name of the database table they should be
stored in. In short, less hard-coded column names and table names in
SQL, more inspection of tags and a special interface.