Remove sqlite import.
The sqlite3 import was included into the reflect.go file so it would be
downloaded automatically with go get, and tests would run with no
further downloads.

But now go get has a -t flag that downloads the files for the test
packages, too, so this is unnecessary.

Worse, because it uses CGo, the sqlite3 import added roughly 20 seconds
to the build time for any package that imports it. Ew.

So the solution was simple: remove the import.