When querying in test, include expressions.
We used `*` in a select query, but were unmarshalling it to a type that
had a property defined by an expression. This caused the unmarshalling
code to complain, because the result used to populate that property
wasn't available, so we had a mismatch and it didn't know what to do.

The solution was to use GetQuotedFieldsAndExpressions to query for the
fields we want, which is what we should have been doing in the first
place.