Update our license structure.
We're now declaring copyright belongs to the pan authors, and adding an
authors file showing the list of people who have a copyright claim.

My understanding is that DramaFever would like to opt-out of this
structure, and so I've left them in their spot in the LICENSE file,
instead of moving their claim to the AUTHORS file.