Implement Query.String.
Implement the String method for query, to return the SQL (with values
injected) that would be executed on the server. This is not quite
perfect yet--we're using the fmt.Sprintf `%v` output for each arg right
now, instead of the actual SQL syntax, but this should be good enough to
give people an idea when debugging. We could improve on this by, e.g.,
quoting strings. I'd like, one day, to have that output be
copy/paste-able valid SQL, but, as we say to death, not today.

I also added a benchmark for the String function, and renamed our
current benchmarks, to make room for future benchmarking.