Fix column output, make our caching parallel-safe.
Make all our tests run in parallel.

Add a RWMutex to our struct reflection cache. This removes annoying race
conditions we had in our caching, without a noticeable performance
impact.

I also added a column cache, so the Column function could only use
reflection the first time it saw a type/property combo, but there was no
measurable performance benefit, so I dropped it.

I changed readStruct to only populate the values return variable if the
needsValues argument is set to true. This has a probably immeasurable
performance impact, but more importantly makes behaviour consistent,
regardless of whether values are retrieved from the cache or from
reflect directly.

I dropped the getColumn helper, which Column called through to, as
Column was the only thing calling it. So now it's just Column. I also
fixed the bug that had us returning the bare column name, instead of the
absolute column name, as the rest of the library does. This required an
update to our sqlTable tests.