Use sql.Rows' Columns method when unmarshaling.
Take advantage of having the Column names available to us to do a better
job of matching columns to properties.

Drop docstrings, which are all out of date and need rewritten.

Rename VariableList to Placeholders, which is really what it's
generating.

Update our Unmarshal test to use Postgres, not sqlite, which doesn't
support table.column_name syntax, and was giving me trouble. Eventually,
we should find a fix for this?