Add docs, return ErrNeedsFlush.
Add docstrings so golint is happy.

Explicitly ignore the output of our String function when benchmarking,
to make go vet happy.

Add an ErrNeedsFlush error, which is returned when a Query is used but
has dangling expressions that weren’t properly flushed.

Add a test that ErrNeedsFlush is returned, as expected.