Upgrade to Go 1.18, add SQLiteString.
Upgrade to Go 1.18, adding a go.mod and go.sum. Change all instances of
`interface{}` to `any`.

Update `Insert` to use generics so we no longer need to convert to a
`[]pan.SQLTableNamer` before passing slices in, slices of any type that
fills the `pan.SQLTableNamer` interface will now work. Should be
backwards-compatible. It also has the benefit of enforcing that all
values need to be of the same type, which was implied by the comment
prior to this, so I'm not considering it a breaking change.

Add an SQLiteString method to Query, returning a string that can be used
for SQLite. Technically, this is the same as MySQLString. But it feels
weird using MySQLString in SQLite code, and now if we discover any
special handling we need to do for SQLite, we can just put it in this
method.

🎵 Now Playing: Better Than That
🎵 Artist: Sub-Radio
🎵 Album: Better Than That